pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"
[project]
name = "trunk-yls"
version = "0.0.4"
authors = [
  { name="Kilian Tyler", email="trunk-yls@ktyler.me" },
]
description = "Personal yaml-language-server check"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
  "ruamel.yaml",
]
[project.urls]
Homepage = "https://github.com/kiliantyler/trunk-yls"
Issues = "https://github.com/kiliantyler/trunk-yls/issues"

[project.scripts]
trunk-yls = "trunk_yls.yls_check:main"

[tool.setuptools]
packages = ["trunk_yls"]

[tool.setuptools.package-dir]
trunk_yls = "src/trunk_yls"